Change signature verification procedures to return 0 upon success, and non-zero upon failure, consistent with safe_memcmp

+++ include/crypto/sign/uov/uov.h
@@ -42,7 +42,7 @@ void uov_gen_key (uint8_t * p1_out, uint8_t * p2_out, uint8_t * p3_out, uint8_t 
 
 uint8_t uov_sign (const uint8_t * seed_sk, const uint8_t * o, const uint8_t * p1, const uint8_t * s, const uint8_t * msg, size_t len, const uint8_t * salt, uint8_t * out);
 
-uint8_t uov_verify (const uint8_t * p1, const uint8_t * p2, const uint8_t * p3, const uint8_t * msg, size_t len, const uint8_t * salt, const uint8_t * sig);
+uint64_t uov_verify (const uint8_t * p1, const uint8_t * p2, const uint8_t * p3, const uint8_t * msg, size_t len, const uint8_t * salt, const uint8_t * sig);
 
 #ifdef __cplusplus
 }
